NFC: Move DL functions to CAPI
